Fixed a bug where values were incorrectly parsed Changed float to double and added messages

+++ HistoricalData.h
@@ -7,8 +7,8 @@
 
 struct Tick {
     long timestamp;
-    float price;
-    float volume;
+    double price;
+    double volume;
 };
 
 class HistoricalData {
